Crate plutus

Minimal Plutus data types and encoding implementation

This crate implements the Data type used by Untyped Plutus Core, a low-level language used by Cardano smart contracts, along with traits for encoding Rust types as Plutus data and Plutus data CBOR bytes.

This crate is developed as part of the Partner Chains toolkit for use in no_std contexts where existing alternatives (uplc, cardano-serialization-lib) can not be used, and as such is not intended to be feature-complete for general use.

Structs

MapDatumEntry

Key-value pair stored in Datum::MapDatum

Enums

Datum

Plutus datum type

Traits

ToDatum

Trait for types that can be encoded as a Plutus Datum.

Functions

to_datum_cbor_bytes

Encodes a message as a Plutus Datum and returns CBOR encoding of this datum.